requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pdf2image>=1.16.0
Pillow>=9.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
